requirements.txt
Faker==1.0.4
ipaddress==1.0.22
PyMySQL==0.9.3
python-dateutil==2.8.0
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.1
text-unidecode==1.2
